add:完善管理员审核功能

+++ src/main/java/com/example/healthdb/model/request/RechargeRequest.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class RechargeRequest {
     private Integer id;
-    private Double money;
+    private Float money;
 }
